+++ examples/resources/langfuse_evaluation_rule/import.sh
@@ -1,0 +1,2 @@
+# Import ID format: <project_public_key>:<project_secret_key>:<evaluation_rule_id>
+terraform import langfuse_evaluation_rule.production_generations "pk-lf-...:sk-lf-...:cm1rule000000000000000000"