+++ examples/resources/langfuse_evaluator/import.sh
@@ -1,0 +1,2 @@
+# Import ID format: <project_public_key>:<project_secret_key>:<evaluator_id>
+terraform import langfuse_evaluator.helpfulness "pk-lf-...:sk-lf-...:cm1evaluator000000000000000"