+++ internal/langfuse/mocks/mock_client_factory.go
@@ -9,6 +9,7 @@ type mockClientFactory struct {
 	AdminClient          *MockAdminClient
 	OrganizationClient   *MockOrganizationClient
 	LlmConnectionsClient *MockLlmConnectionsClient
+	EvaluatorsClient     *MockEvaluatorsClient
 }
 
 func NewMockClientFactory(ctrl *gomock.Controller) *mockClientFactory {
@@ -16,6 +17,7 @@ func NewMockClientFactory(ctrl *gomock.Controller) *mockClientFactory {
 		AdminClient:          NewMockAdminClient(ctrl),
 		OrganizationClient:   NewMockOrganizationClient(ctrl),
 		LlmConnectionsClient: NewMockLlmConnectionsClient(ctrl),
+		EvaluatorsClient:     NewMockEvaluatorsClient(ctrl),
 	}
 }
 
@@ -29,4 +31,8 @@ func (cf *mockClientFactory) NewOrganizationClient(publicKey, privateKey string)
 
 func (cf *mockClientFactory) NewLlmConnectionsClient(publicKey, privateKey string) langfuse.LlmConnectionsClient {
 	return cf.LlmConnectionsClient
+}
+
+func (cf *mockClientFactory) NewEvaluatorsClient(publicKey, secretKey string) langfuse.EvaluatorsClient {
+	return cf.EvaluatorsClient
 }